present more challenge metadata for non-http challenge types
When -t is used to perform a challenge using dns-01 or some other
hypothetical challenge type, only the token and thumbprint are
displayed.  That's not enough information to set up challenge response
manually.  This patch extends the chngproc IPC semantics and the
stdout from netproc to provide the challenge type and domain name
being challenged.  That's enough for external challenge setup.

This particularly comes into interest when retrieving a single
certificate with multiple subjectAlternativeNames, especially when those
alt names are in different DNS domains and might require different
account authorizations to perform updates.